Fix: fix tests
